removed curious2.js in favor of curious.js, and add default post filter for axios
