+++ make_doc
@@ -1,3 +1,3 @@
 #!/bin/bash
 mkdir -p doc
-jsdoc --destination doc curious2.js "$@"
+jsdoc --destination doc curious.js "$@"
