+++ test
@@ -1,0 +1,1 @@
+NODE_PATH=lib:. node_modules/mocha/bin/mocha tests
